requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.7.22
cffi==1.16.0
kafka-python==2.0.2
minio==7.2.0
py4j==0.10.9.7
pycparser==2.21
pycryptodome==3.19.0
pyspark==3.5.0
PyYAML==6.0.1
urllib3==2.0.7
